Refactors sidebar navigation
Introduces a `SidebarNavLink` component to streamline the creation of navigation links within the admin sidebar.

This change improves code readability and maintainability by encapsulating the logic for creating `NavLink` components with associated `SidebarMenuButton` components. It reduces repetition and makes it easier to add or modify sidebar links in the future.